usm: tests: classification: Move redis classification test to USM linux file
