+++ pkg/network/tracer/tracer_usm_linux_test.go
@@ -22,6 +22,7 @@ import (
 	"testing"
 	"time"
 
+	redis2 "github.com/go-redis/redis/v9"
 	gorilla "github.com/gorilla/mux"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -42,6 +43,7 @@ import (
 	protocolsmongo "github.com/DataDog/datadog-agent/pkg/network/protocols/mongo"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/mysql"
 	pgutils "github.com/DataDog/datadog-agent/pkg/network/protocols/postgres"
+	"github.com/DataDog/datadog-agent/pkg/network/protocols/redis"
 	prototls "github.com/DataDog/datadog-agent/pkg/network/protocols/tls/openssl"
 	"github.com/DataDog/datadog-agent/pkg/network/tracer/connection"
 	"github.com/DataDog/datadog-agent/pkg/network/tracer/connection/kprobe"
@@ -54,6 +56,7 @@ const (
 	mysqlPort    = "3306"
 	postgresPort = "5432"
 	mongoPort    = "27017"
+	redisPort    = "6379"
 	kafkaPort    = "9092"
 
 	produceAPIKey = 0
@@ -1436,6 +1439,161 @@ func testMongoProtocolClassification(t *testing.T, tr *Tracer, clientHost, targe
 			},
 			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Mongo}),
 			teardown:   mongoTeardown,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			testProtocolClassificationInner(t, tt, tr)
+		})
+	}
+}
+
+func testRedisProtocolClassification(t *testing.T, tr *Tracer, clientHost, targetHost, serverHost string) {
+	skipFunc := composeSkips(skipIfNotLinux, skipIfUsingNAT)
+	skipFunc(t, testContext{
+		serverAddress: serverHost,
+		serverPort:    redisPort,
+		targetAddress: targetHost,
+	})
+
+	defaultDialer := &net.Dialer{
+		LocalAddr: &net.TCPAddr{
+			IP: net.ParseIP(clientHost),
+		},
+	}
+
+	redisTeardown := func(t *testing.T, ctx testContext) {
+		redis.NewClient(ctx.serverAddress, defaultDialer)
+		client := ctx.extras["client"].(*redis2.Client)
+		timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+		defer cancel()
+		require.NoError(t, client.FlushDB(timedContext).Err())
+	}
+
+	// Setting one instance of redis server for all tests.
+	serverAddress := net.JoinHostPort(serverHost, redisPort)
+	targetAddress := net.JoinHostPort(targetHost, redisPort)
+	require.NoError(t, redis.RunServer(t, serverHost, redisPort))
+
+	tests := []protocolClassificationAttributes{
+		{
+			name: "set",
+			context: testContext{
+				serverPort:    redisPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client := redis.NewClient(ctx.targetAddress, defaultDialer)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				client.Ping(timedContext)
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*redis2.Client)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				client.Set(timedContext, "key", "value", time.Minute)
+			},
+			teardown:   redisTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Redis}),
+		},
+		{
+			name: "get",
+			context: testContext{
+				serverPort:    redisPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client := redis.NewClient(ctx.targetAddress, defaultDialer)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				client.Set(timedContext, "key", "value", time.Minute)
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*redis2.Client)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				res := client.Get(timedContext, "key")
+				val, err := res.Result()
+				require.NoError(t, err)
+				require.Equal(t, "value", val)
+			},
+			teardown:   redisTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Redis}),
+		},
+		{
+			name: "get unknown key",
+			context: testContext{
+				serverPort:    redisPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client := redis.NewClient(ctx.targetAddress, defaultDialer)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				client.Ping(timedContext)
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*redis2.Client)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				res := client.Get(timedContext, "unknown")
+				require.Error(t, res.Err())
+			},
+			teardown:   redisTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Redis}),
+		},
+		{
+			name: "err response",
+			context: testContext{
+				serverPort:    redisPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				conn, err := defaultDialer.DialContext(timedContext, "tcp", ctx.targetAddress)
+				require.NoError(t, err)
+				_, err = conn.Write([]byte("+dummy\r\n"))
+				require.NoError(t, err)
+			},
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Redis}),
+		},
+		{
+			name: "client id",
+			context: testContext{
+				serverPort:    redisPort,
+				serverAddress: serverAddress,
+				targetAddress: targetAddress,
+				extras:        make(map[string]interface{}),
+			},
+			preTracerSetup: func(t *testing.T, ctx testContext) {
+				client := redis.NewClient(ctx.targetAddress, defaultDialer)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				client.Ping(timedContext)
+				ctx.extras["client"] = client
+			},
+			postTracerSetup: func(t *testing.T, ctx testContext) {
+				client := ctx.extras["client"].(*redis2.Client)
+				timedContext, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+				defer cancel()
+				res := client.ClientID(timedContext)
+				require.NoError(t, res.Err())
+			},
+			teardown:   redisTeardown,
+			validation: validateProtocolConnection(&protocols.Stack{Application: protocols.Redis}),
 		},
 	}
 	for _, tt := range tests {
